upcoming Facebook events info is loaded to recyclerview

+++ app/src/main/java/Entities/FBEvent.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class FBEvent {
 
     public String Name;
-    public int  event_id;
+    public long  event_id;
     public String description;
     public String start_time;
 
@@ -17,7 +17,7 @@ public class FBEvent {
         // Default constructor required for calls to DataSnapshot.getValue(User.class)
     }
 
-    public FBEvent(String name, String description, String start_time,int event_id) {
+    public FBEvent(String name, String description, String start_time,long event_id) {
         this.Name = name;
         this.event_id= event_id;
         this.description= description;
